fix - quoting from m/l: I've already posted about this, but will explain in length for everyone's benefit.
This is because the koffice-libs ebuild is bad. For the packaged makefiles 
feature to work, the makefiles (Makefile.am in this case) must _not_ be 
changed manually after kde-meta_src_unpack. That's because such a change 
doesn't affect the makefiles tarball. And when using the packaged makefiles, 
this caused Makefile.am to be newer than the packaged Makefile.in, so it 
tried to regen Makefile.in, which doesn't work unless you do a full make -f 
admin/Makefile.common (and I don't know why).